move tree-option stuff to /option
