+++ packages/ng/libraries/core/src/lib/index.ts
@@ -1,1 +1,2 @@
 export * from './translate/index';
+export * from './tree/index';
