+++ packages/ng/libraries/core/src/lib/tree/index.ts
@@ -1,0 +1,1 @@
+export * from './tree.model';
