+++ packages/ng/libraries/core/src/lib/tree/tree.model.ts
@@ -1,4 +1,4 @@
 export interface ILuTree<T = any> {
 	value: T;
-	children: ILuTree<T>[]
+	children: ILuTree<T>[];
 }
