+++ packages/ng/libraries/option/src/lib/item/index.ts
@@ -1,3 +1,10 @@
 export * from './option-item.component';
 export * from './option-item.model';
 export * from './option-item.module';
+
+export * from './tree-option-item.component';
+export * from './tree-option-item.model';
+export * from './tree-option-item.module';
+export * from './tree-option-item.token';
+export * from './tree-option-item.translate';
+export * from './tree-option-item.intl';
