+++ packages/ng/libraries/option/src/lib/item/tree-option-item.component.scss
