+++ packages/ng/libraries/option/src/lib/item/tree-option-item.intl.ts
@@ -1,7 +1,7 @@
 import { Injectable, LOCALE_ID, Inject } from '@angular/core';
 import { LU_TREE_OPTION_ITEM_TRANSLATIONS } from './tree-option-item.token';
 import { ILuTreeOptionItemLabel } from './tree-option-item.translate';
-import { ALuIntl } from '../../../translate/index';
+import { ALuIntl } from '@lucca-front/ng/core';
 
 @Injectable()
 export class LuTreeOptionItemIntl extends ALuIntl<ILuTreeOptionItemLabel> {
