+++ packages/ng/libraries/option/src/lib/item/tree-option-item.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { LuTreeOptionItemComponent } from './tree-option-item.component';
-import { LuTooltipModule } from '../../../overlay/index';
+import { LuTooltipModule } from '@lucca-front/ng/tooltip';
 import { LuTreeOptionItemIntl } from './tree-option-item.intl';
 import { luTreeOptionItemTranslations } from './tree-option-item.translate';
 import { LU_TREE_OPTION_ITEM_TRANSLATIONS } from './tree-option-item.token';
@@ -23,7 +23,6 @@ import { CommonModule } from '@angular/common';
 	providers: [
 		LuTreeOptionItemIntl,
 		{ provide: LU_TREE_OPTION_ITEM_TRANSLATIONS, useValue: luTreeOptionItemTranslations },
-
 	],
 })
 export class LuTreeOptionItemModule {}
