+++ packages/ng/libraries/option/src/lib/operator/feeder/index.ts
@@ -1,2 +1,5 @@
 export * from './option-feeder.component';
 export * from './option-feeder.module';
+
+export * from './tree-option-feeder.module';
+export * from './tree-option-feeder.component';
