+++ packages/ng/libraries/option/src/lib/operator/feeder/tree-option-feeder.component.ts
@@ -1,7 +1,7 @@
 import { ChangeDetectionStrategy, Component, forwardRef, Input } from '@angular/core';
 import { BehaviorSubject } from 'rxjs';
 import { ILuTreeOptionOperator, ALuTreeOptionOperator } from '../tree-option-operator.model';
-import { ILuTree } from '../../../tree.model';
+import { ILuTree } from '@lucca-front/ng/core';
 
 @Component({
 	selector: 'lu-tree-option-feeder',
