+++ packages/ng/libraries/option/src/lib/operator/for-options/index.ts
@@ -1,2 +1,5 @@
 export * from './for-options.directive';
 export * from './for-options.module';
+
+export * from './for-tree-options.directive';
+export * from './for-tree-options.module';
