+++ packages/ng/libraries/option/src/lib/operator/index.ts
@@ -2,5 +2,9 @@ export * from './feeder/index';
 export * from './pager/index';
 export * from './searcher/index';
 export * from './for-options/index';
+
 export * from './option-operator.module';
 export * from './option-operator.model';
+
+export * from './tree-option-operator.model';
+export * from './tree-option-operator.module';
