+++ packages/ng/libraries/option/src/lib/operator/pager/index.ts
@@ -1,2 +1,5 @@
 export * from './option-pager.component';
 export * from './option-pager.module';
+
+export * from './tree-option-pager.component';
+export * from './tree-option-pager.module';
