+++ packages/ng/libraries/option/src/lib/operator/searcher/index.ts
@@ -1,2 +1,5 @@
 export * from './option-searcher.component';
 export * from './option-searcher.module';
+
+export * from './tree-option-searcher.module';
+export * from './tree-option-searcher.component';
