+++ packages/ng/libraries/option/src/lib/operator/tree-option-operator.module.ts
@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { LuTreeOptionFeederModule } from './feeder/index';
-import { LuForTreeOptionsModule } from './for-tree-options/index';
+import { LuForTreeOptionsModule } from './for-options/index';
 import { LuTreeOptionSearcherModule } from './searcher/index';
 import { LuTreeOptionPagerModule } from './pager/index';
