+++ packages/ng/libraries/option/src/lib/picker/index.ts
@@ -2,3 +2,8 @@ export * from './option-picker.module';
 export * from './option-picker.component';
 export * from './option-picker-advanced.component';
 export * from './option-picker.model';
+
+export * from './tree-option-picker.module';
+export * from './tree-option-picker.model';
+export * from './tree-option-picker.component';
+export * from './tree-option-picker-advanced.component';
