+++ packages/ng/libraries/option/src/lib/picker/tree-option-picker.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from '@angular/core';
 import { LuTreeOptionPickerComponent } from './tree-option-picker.component';
 import { CommonModule } from '@angular/common';
 import { OverlayModule } from '@angular/cdk/overlay';
-import { LuScrollModule } from '../../../scroll/index';
+import { LuScrollModule } from '@lucca-front/ng/scroll';
 import { A11yModule } from '@angular/cdk/a11y';
 import { LuTreeOptionPickerAdvancedComponent } from './tree-option-picker-advanced.component';
