+++ packages/ng/libraries/option/src/lib/selector/all/index.ts
@@ -3,3 +3,6 @@ export * from './select-all.component';
 export * from './select-all.intl';
 export * from './select-all.token';
 export * from './select-all.translate';
+
+export * from './tree-select-all.module';
+export * from './tree-select-all.component';
