+++ packages/ng/libraries/option/src/lib/selector/index.ts
@@ -1,4 +1,7 @@
 export * from './option-selector.model';
 export * from './option-selector.module';
 
+export * from './tree-option-selector.model';
+export * from './tree-option-selector.module';
+
 export * from './all/index';
