+++ packages/ng/libraries/option/src/lib/selector/tree-option-selector.model.ts
@@ -1,4 +1,4 @@
-import { ILuOptionSelector, ALuOptionSelector } from '../../../option/index';
+import { ILuOptionSelector, ALuOptionSelector } from './option-selector.model';
 
 export interface ILuTreeOptionSelector<T = any> extends ILuOptionSelector {}
 export abstract class ALuTreeOptionSelector<T = any> extends ALuOptionSelector implements ILuTreeOptionSelector<T> {}
